A Book Shop Management System
This project revolves around the development and implementation of a database management system tailored for a bookshop environment. The system aims to centralize and organize various datasets pertaining to books, authors, sales, ratings, and other relevant information. By leveraging the capabilities of a relational DBMS, we aim to create a cohesive platform that facilitates efficient data management, querying, and reporting. Through this project, we seek to empower bookshop owners and managers with the tools necessary to make informed decisions and optimize business operations.